+++ src/Symfony/Component/RateLimiter/Policy/TokenBucketLimiter.php
@@ -81,14 +81,14 @@ final class TokenBucketLimiter implements LimiterInterface
                     $retryAfter += $this->rate->calculateTimeForTokens(1);
                 }
 
-                $reservation = new Reservation($now, new RateLimit($bucket->getAvailableTokens($now), \DateTimeImmutable::createFromFormat('U', floor($retryAfter)), true, $this->maxBurst));
+                $reservation = new Reservation($now, new RateLimit($bucket->getAvailableTokens($now), \DateTimeImmutable::createFromTimestamp((int) $retryAfter), true, $this->maxBurst, $this->resetAt($availableTokens - $tokens, $now)));
             } else {
                 $remainingTokens = $tokens - $availableTokens;
                 $waitDuration = $this->rate->calculateTimeForTokens($remainingTokens);
 
                 if (null !== $maxTime && $waitDuration > $maxTime) {
                     // process needs to wait longer than set interval
-                    $rateLimit = new RateLimit($availableTokens, \DateTimeImmutable::createFromFormat('U', floor($now + $waitDuration)), false, $this->maxBurst);
+                    $rateLimit = new RateLimit($availableTokens, \DateTimeImmutable::createFromTimestamp((int) ($now + $waitDuration)), false, $this->maxBurst, $this->resetAt($availableTokens, $now));
 
                     throw new MaxWaitDurationExceededException(\sprintf('The rate limiter wait time ("%d" seconds) is longer than the provided maximum time ("%d" seconds).', $waitDuration, $maxTime), $rateLimit);
                 }
@@ -97,7 +97,7 @@ final class TokenBucketLimiter implements LimiterInterface
                 // so no tokens are left for other processes.
                 $bucket->setTokens($availableTokens - $tokens);
 
-                $reservation = new Reservation($now + $waitDuration, new RateLimit(0, \DateTimeImmutable::createFromFormat('U', floor($now + $waitDuration)), false, $this->maxBurst));
+                $reservation = new Reservation($now + $waitDuration, new RateLimit(0, \DateTimeImmutable::createFromTimestamp((int) ($now + $waitDuration)), false, $this->maxBurst, $this->resetAt($availableTokens - $tokens, $now)));
             }
 
             if (0 !== $tokens) {
@@ -117,5 +117,10 @@ final class TokenBucketLimiter implements LimiterInterface
         } catch (MaxWaitDurationExceededException $e) {
             return $e->getRateLimit();
         }
+    }
+
+    private function resetAt(int $availableTokens, float $now): \DateTimeImmutable
+    {
+        return \DateTimeImmutable::createFromTimestamp((int) ($now + $this->rate->calculateTimeForTokens(max(0, $this->maxBurst - $availableTokens))));
     }
 }